+++ R/retrieve.R
@@ -90,7 +90,7 @@ renv_retrieve_impl <- function(package) {
     # if the requested record already exists in the cache,
     # we'll use that package for install
     cacheable <-
-      settings$use.cache(project = state$project) &&
+      renv_cache_config_enabled(project = state$project) &&
       renv_record_cacheable(record)
 
     if (cacheable) {